# backend/bin/Debug/net8.0/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Art Gallery",
      "description": "Backend service that provides resources for the beautiful art gallery.",
      "contact": {
        "name": "Tanay and Jeetasha",
        "email": "tanay4847.be22@chitkara.edu.in"
      }
    },
    "paths": {
...
      "/api/users": {
        "get": {
          "tags": [
            "User"
          ],
          "summary": "Retrieves all users.",
          "description": "Sample request:\r\n            \r\nGET /api/users",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "User"
          ],
          "summary": "Adds a new user.",
          "description": "Sample request:\r\n            \r\nPOST /api/users\r\n{\r\n    \"firstName\": \"John\",\r\n    \"lastName\": \"Doe\",\r\n    \"description\": \"User description\",\r\n    \"role\": \"User\",\r\n}",
          "requestBody": {
            "description": "The user to add",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Created",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "/api/users/{id}": {
        "get": {
          "tags": [
            "User"
          ],
          "summary": "Retrieves a specific user by its ID.",
          "description": "Sample request:\r\n            \r\nGET /api/users/1",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the user to retrieve",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "User"
          ],
          "summary": "Updates an existing user.",
          "description": "Sample request:\r\n\r\nPUT /api/users/1\r\n{\r\n    \"id\": 1,\r\n    \"firstName\": \"John\",\r\n    \"lastName\": \"Doe\",\r\n    \"description\": \"Updated description\",\r\n    \"role\": \"User\"\r\n}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the user to update.",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "description": "The updated user information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "User"
          ],
          "summary": "Deletes a user.",
          "description": "Sample request:\r\n\r\nDELETE /api/users/1",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the user to delete",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
      "/api/users/{id}/credentials": {
        "patch": {
          "tags": [
            "User"
          ],
          "summary": "Updates the credentials of a user.",
          "description": "Sample request:\r\n            \r\nPATCH /api/users/1/credentials\r\n{\r\n    \"email\": \"user@example.com\",\r\n    \"password\": \"new_password\"\r\n}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the user to update",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "description": "Updated login credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Login"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Login"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Login"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "No Content"
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "Login": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "nullable": true
            },
            "password": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "ProblemDetails": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "nullable": true
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "status": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "detail": {
              "type": "string",
              "nullable": true
            },
            "instance": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": {
  
          }
        },
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "email": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "lastName": {
              "type": "string",
              "nullable": true
            },
            "passwordHash": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "role": {
              "type": "string",
              "nullable": true
            },
            "createdDate": {
              "type": "string",
              "format": "date-time"
            },
            "modifiedDate": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        }
      }
    }
  }